+++ code/parachain/frame/composable-traits/src/cosmos.rs
@@ -1,0 +1,15 @@
+
+// Hash creates a new address from address type and key.
+// The functions should only be used by new types defining their own address function
+// (eg public keys).
+/// https://github.com/cosmos/cosmos-sdk/blob/main/types/address/hash.go
+pub fn addess_hash(typ : &str, key: &[u8])  -> [u8; 32] {
+	use sha2::{Sha256, Digest};
+	let mut hasher = Sha256::default();
+	hasher.update(typ.as_bytes());
+	let th = hasher.finalize();
+	let mut hasher = Sha256::default();
+	hasher.update(th);
+	hasher.update(key);	
+	hasher.finalize().into()
+}
