+++ code/parachain/frame/pallet-multihop-xcm-ibc/src/lib.rs
@@ -35,7 +35,8 @@ pub mod pallet {
 
 	use composable_traits::{
 		prelude::ToString,
-		xcm::memo::{ChainInfo, Forward, MemoData},
+		xcm::memo::{ChainInfo},
+		ibc::{Forward, MemoData},
 	};
 	use sp_std::boxed::Box;
 
@@ -175,7 +176,7 @@ pub mod pallet {
 				if let Some(memo_memo) = last_memo_data {
 					forward.next = Some(Box::new(memo_memo));
 				};
-				let new_memo = MemoData::new(forward);
+				let new_memo = MemoData::forward(forward);
 				last_memo_data = Some(new_memo);
 			}
 			Ok(last_memo_data)
